requirements.txt
atlassian-python-api
PyYAML
PyGithub==1.51
pytz==2019.1